no message
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@665 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ trunk/rdo_lib/rdo_runtime/rdofunc.h
@@ -42,7 +42,7 @@ class RandGeneratorUniform: public RandGenerator
 {
 public:
 	RandGeneratorUniform( long int _seed = 123456789 ): RandGenerator( _seed ) {}
-	double next( int from, int to ) {
+	double next( double from, double to ) {
 		return u01() * ( to - from ) + from;
 	}
 };
